requirements_simple.txt
# 简化的后端API依赖列表
# 只包含核心功能所需的依赖

# Web框架
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# LlamaIndex 核心
llama-index-core>=0.14.0
llama-index-llms-openai>=0.4.0
llama-index-embeddings-openai>=0.5.0
llama-index-readers-file>=0.5.0

# 向量数据库
chromadb>=1.1.0
llama-index-vector-stores-chroma>=0.5.0

# 数据处理
pandas>=2.2.0
numpy>=1.26.0

# PDF处理
pdfplumber>=0.10.0

# AI和ML
openai>=1.99.0

# 工具库
python-dotenv>=1.0.0
pydantic>=2.11.0

# 可选：如果需要表格提取增强功能
# openpyxl>=3.1.0
# tabula-py>=2.8.0
